bugfix for category controller and change category tests

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -67,6 +67,11 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if (count($category->products) != 0) {
+            return response('Category '.$category->id.' has products', 409)
+                  ->header('Content-Type', 'text/plain');
+        }
+
         $category->delete();
 
         return response()->noContent();
